@@ -5,17 +5,16 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9, <4.0"
+requires-python = ">=3.10"
 dependencies = [
-    "langchain-core<1.0.0,>=0.3.66",
-    "langchain-text-splitters<1.0.0,>=0.3.8",
+    "langchain-core<2.0.0,>=0.3.75",
+    "langchain-text-splitters<1.0.0,>=0.3.9",
     "langgraph>=0.6.0",
     "pydantic>=2.7.4",
 ]
 
-
 name = "langchain"
-version = "1.0.0dev1"
+version = "1.0.0a2"
 description = "Building applications with LLMs through composability"
 readme = "README.md"
 
@@ -46,23 +45,22 @@ repository = "https://github.com/langchain-ai/langchain"
 [dependency-groups]
 test = [
     "pytest<9,>=8",
-    "pytest-cov<5.0.0,>=4.0.0",
-    "pytest-watcher<1.0.0,>=0.2.6",
-    "pytest-asyncio<1.0.0,>=0.23.2",
-    "pytest-socket<1.0.0,>=0.6.0",
-    "syrupy<5.0.0,>=4.0.2",
-    "pytest-xdist<4.0.0,>=3.6.1",
-    "blockbuster<1.6,>=1.5.18",
+    "pytest-cov>=4.0.0",
+    "pytest-watcher>=0.2.6",
+    "pytest-asyncio>=0.23.2",
+    "pytest-socket>=0.6.0",
+    "syrupy>=4.0.2",
+    "pytest-xdist>=3.6.1",
     "langchain-tests",
-    "langchain-core",
     "langchain-text-splitters",
     "langchain-openai",
     "toml>=0.10.2",
+    "pytest-mock"
 ]
 codespell = ["codespell<3.0.0,>=2.2.0"]
 lint = [
-    "ruff<0.13,>=0.12.2",
-    "mypy<1.16,>=1.15",
+    "ruff>=0.12.2",
+    "mypy>=1.15",
 ]
 typing = [
     "types-toml>=0.10.8.20240310",
@@ -70,11 +68,10 @@ typing = [
 
 test_integration = [
     "vcrpy>=7.0",
-    "urllib3<2; python_version < \"3.10\"",
-    "wrapt<2.0.0,>=1.15.0",
-    "python-dotenv<2.0.0,>=1.0.0",
-    "cassio<1.0.0,>=0.1.0",
-    "langchainhub<1.0.0,>=0.1.16",
+    "wrapt>=1.15.0",
+    "python-dotenv>=1.0.0",
+    "cassio>=0.1.0",
+    "langchainhub>=0.1.16",
     "langchain-core",
     "langchain-text-splitters",
 ]
@@ -86,8 +83,9 @@ langchain-text-splitters = { path = "../text-splitters", editable = true }
 langchain-openai = { path = "../partners/openai", editable = true }
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 exclude = ["tests/integration_tests/examples/non-utf8-encoding.py"]
+line-length = 100
 
 [tool.mypy]
 strict = "True"
@@ -95,6 +93,7 @@ strict_bytes = "True"
 ignore_missing_imports = "True"
 enable_error_code = "deprecated"
 report_deprecated_as_note = "True"
+exclude = ["tests/unit_tests/agents/*", "tests/integration_tests/agents/*"]
 
 # TODO: activate for 'strict' checking
 disallow_untyped_calls = "False"
@@ -136,13 +135,23 @@ flake8-annotations.allow-star-arg-any = true
   "D", # Documentation rules
   "PLC0415", # Imports should be at the top. Not always desirable for tests
 ]
+"langchain/agents/*" = [
+    "ANN401", # we use Any right now, need to narrow
+    "E501", # line too long, needs to fix
+    "A002", # input is shadowing builtin
+    "A001", # input is shadowing builtin
+    "B904", # use from for exceptions
+    "PLR2004", # magic values are fine for this case
+    "C901", # too complex
+    "TRY004", # type error exception
+    "PLR0912", # too many branches
+    "PLR0911", # too many return statements
+]
+"tests/unit_tests/agents/*" = ["ALL"]
+"tests/integration_tests/agents/*" = ["ALL"]
 
 [tool.ruff.lint.extend-per-file-ignores]
 "scripts/check_imports.py" = ["ALL"]
-
-"langchain/globals.py" = [
-    "PLW"
-]
 
 "langchain/chat_models/base.py" = [
     "ANN",
